+++ src-tauri/src/lib.rs
@@ -198,6 +198,7 @@ pub fn run() {
             storage_commands::media_commands::llm_stream_cancel,
             storage_commands::media_commands::llm_list_models,
             storage_commands::mari_commands::professor_mari_prompt,
+            storage_commands::mari_commands::professor_mari_apply_staged_changes,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
